requirements.txt
# Flask Web Framework
Flask==3.0.0
flask-cors==4.0.0

# MongoDB Driver
pymongo==4.6.1

# OpenAI API Client
openai==1.6.1
httpx==0.27.2

# Environment Variable Management
python-dotenv==1.0.0

# Production WSGI Server
gunicorn==21.2.0
